requirements.txt
pywin32>=305
Pillow>=10.0.0
pyinstaller>=6.0.0